+++ src/client/list_deployments.rs
@@ -1,0 +1,45 @@
+use bollard::query_parameters::ListContainersOptionsBuilder;
+use maplit::hashmap;
+
+use crate::{
+    client::Client,
+    docker::{DockerInspectContainer, DockerListContainers},
+    models::{Deployment, LOCAL_DEPLOYMENT_LABEL_KEY, LOCAL_DEPLOYMENT_LABEL_VALUE},
+};
+
+use super::GetDeploymentError;
+
+impl<D: DockerListContainers + DockerInspectContainer> Client<D> {
+    /// Lists all local Atlas deployments.
+    pub async fn list_deployments(&self) -> Result<Vec<Deployment>, GetDeploymentError> {
+        // Build the list containers options which will filter for containers with the local deployment label
+        let list_container_options = ListContainersOptionsBuilder::default()
+            .all(true)
+            .filters(&hashmap! {
+                "label" => vec![format!("{}={}", LOCAL_DEPLOYMENT_LABEL_KEY, LOCAL_DEPLOYMENT_LABEL_VALUE)],
+            })
+            .build();
+
+        // Get all the containers using the list containers options
+        let container_summaries = self
+            .docker
+            .list_containers(Some(list_container_options))
+            .await?;
+
+        // Create the output vector used to return the deployments
+        let mut deployments = Vec::with_capacity(container_summaries.len());
+
+        // Iterate over the container summaries and get the deployment details for each container
+        for container_summary in container_summaries {
+            // Get the container ID from the container summary
+            // This should always be present, but it's cleaner to not use unwrap and skip if it's not present
+            if let Some(container_id) = container_summary.id {
+                // Get the deployment details for the container
+                let deployment = self.get_deployment(container_id.as_ref()).await?;
+                deployments.push(deployment);
+            }
+        }
+
+        Ok(deployments)
+    }
+}
